backend/requirements.txt
flask
flask[async]
googlesearch-python
duckduckgo-search
beautifulsoup4
tqdm
requests
dnspython
pymongo
